fix delete item permission logic

+++ backend/src/resolvers/Mutation.js
@@ -45,8 +45,15 @@ const Mutation = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user { id }}`);
     // 2. check if the user own that item, or have th permission to delete it
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermission = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+
+    if(!ownsItem && !hasPermission) {
+      throw new Error("You don't have permission to do that");
+    }
+
     // TODO
     // 3. delete it
     return ctx.db.mutation.deleteItem({ where }, info);
